(PCP-389) Add new connection test
Adding new connection test, that perform a number of consecutive runs
where PCP connections are created in parallel tasks.

Parsing the connection test configuration parameters directly from the
configuration file as a JSON object; validating them with a JSON schema.

Creating new utility functions to inspect the files in the
certificates-dir and retrieve the common names of the endpoints based on
the file names.

Storing the connection test results in file in CSV format; also
displaying them in console.