(PCP-432) Option for randomizing inter-connection pauses
Using the new RNG to determine inter-connection pauses.

Adding a boolean option to the Connection test for randomizing the
interval between successive connections.

Adding an integer option to the same test for defining the seed of the
RNG (exp. dist.) for inter-connection pauses.

Fixing the way the Connection Task pauses; previously, it used to pause
twice.

Fixing the way the standard deviation of TCP timings is reported to
file, as previously it was reported as 10 times greater.

Fixing the way timing stats are appended to the other results to file,
as previously a comma was missing.

Taking into account the random values when determining the timeout of a
given Connection Test run.

Reporting on stdout when the connections are about to be closed.

Updating docs.