(PCP-579) Have acceptance test against the build from CI
The initial commit of acceptance tests were just a placeholder, and they ran against a hardcoded URL to a pcp-test installer.
This commit updates the acceptance setup to test against an environment specified build of pcp-test.
Because pcp-test depends on the libraries of the the version of puppet-agent it was built against, this commit also updates the puppet agent installation
to ensure that it installs the same version of puppet-agent that was used to build pcp-test.